requirements.txt
aiohttp
pandas
json
mysql.connector
pyodbc
python-dotenv
os
asyncio